requirements.txt
mysql-connector-python==8.0.22
RPi.GPIO==0.7.0
